Separated the UI into two tasks: Result and Setup
